+++ CHANGELOG.md
@@ -25,6 +25,7 @@ _This release is scheduled to be released on 2023-07-01._
 - Use node v19 in github workflow (replacing v14)
 - Refactor formatTime into common util function for default modules
 - Refactor some calendar methods into own class and added tests for them
+- Properly convert newsfeed titles
 
 ### Fixed